requirements.txt
opencv-python==4.1.2.30
numpy==1.18.5
Pillow==7.0.0
streamlit==0.60.0
